chore: move GrammarRepository -> Grammar/GrammarRepository
